+++ diffsims/crystallography/__init__.py
@@ -16,8 +16,8 @@
 # You should have received a copy of the GNU General Public License
 # along with diffsims.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Generation of reciprocal lattice points (crystal plane, reflector, g,
-hkl) for a crystal structure.
+"""Generation of reciprocal lattice vectors (crystal plane, reflector,
+g, hkl) for a crystal structure.
 """
 
 from diffsims.crystallography.reciprocal_lattice_point import (
@@ -26,10 +26,12 @@ from diffsims.crystallography.reciprocal_lattice_point import (
     get_highest_hkl,
     get_hkl,
 )
+from diffsims.crystallography.reciprocal_lattice_vector import ReciprocalLatticeVector
 
 __all__ = [
-    "ReciprocalLatticePoint",
     "get_equivalent_hkl",
     "get_highest_hkl",
     "get_hkl",
+    "ReciprocalLatticePoint",
+    "ReciprocalLatticeVector",
 ]